[KYUUBI #1784] Fix float types lose precision
<!--
Thanks for sending a pull request!

Here are some tips for you:
  1. If this is your first time, please read our contributor guidelines: https://kyuubi.readthedocs.io/en/latest/community/contributions.html
  2. If the PR is related to an issue in https://github.com/apache/incubator-kyuubi/issues, add '[KYUUBI #XXXX]' in your PR title, e.g., '[KYUUBI #XXXX] Your PR title ...'.
  3. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP][KYUUBI #XXXX] Your PR title ...'.
-->

### _Why are the changes needed?_
<!--
Please clarify why the changes are needed. For instance,
  1. If you add a feature, you can talk about the use case of it.
  2. If you fix a bug, you can clarify why it is a bug.
-->

close #1784

### _How was this patch tested?_
- [x] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.readthedocs.io/en/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #1790 from cfmcgrady/kyuubi-1784.

Closes #1784

f6c7dae8 [Fu Chen] fix row-based row set
89e72b73 [Fu Chen] fix ut
cef54bc0 [Fu Chen] fix float types lose precision

Authored-by: Fu Chen <cfmcgrady@gmail.com>
(cherry picked from commit 53bbbb71616c6857fe6cbb730956542bd26e57b7)

+++ externals/kyuubi-spark-sql-engine/src/test/scala/org/apache/kyuubi/schema/RowSetSuite.scala
@@ -142,7 +142,7 @@ class RowSetSuite extends KyuubiFunSuite {
     val floatCol = cols.next().getDoubleVal
     floatCol.getValues.asScala.zipWithIndex.foreach {
       case (b, 11) => assert(b === 0)
-      case (b, i) => assert(b === java.lang.Float.valueOf(s"$i.$i"))
+      case (b, i) => assert(b.toFloat === java.lang.Float.valueOf(s"$i.$i"))
     }
 
     val doubleCol = cols.next().getDoubleVal
@@ -226,7 +226,7 @@ class RowSetSuite extends KyuubiFunSuite {
 
     val r4 = iter.next().getColVals
     assert(r4.get(4).getI64Val.getValue === 3)
-    assert(r4.get(5).getDoubleVal.getValue === 3.3f)
+    assert(r4.get(5).getDoubleVal.getValue.toFloat === 3.3f)
 
     val r5 = iter.next().getColVals
     assert(r5.get(6).getDoubleVal.getValue === 4.4d)
